Move Daemon to its own module.
